laptopguardian_New/requirements.txt
dash==2.18.2
plotly==5.24.1
psutil==7.0.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
win10toast==0.9
